Added base user model & publish model on install
