+++ src/Models/User.php
@@ -1,37 +1,10 @@
 <?php
 
-namespace Rapidez\Statamic\Models;
+namespace App\Models;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Foundation\Auth\User as Authenticatable;
-use Illuminate\Notifications\Notifiable;
-use Statamic\Notifications\PasswordReset;
+use Rapidez\Statamic\Models\BaseUser;
 
-class User extends Authenticatable
+class User extends BaseUser
 {
-    use HasFactory, Notifiable;
-
-    protected $fillable = [
-        'name',
-        'email',
-        'password',
-    ];
-
-    protected $hidden = [
-        'password',
-        'remember_token',
-    ];
-
-    protected function casts(): array
-    {
-        return [
-            'email_verified_at' => 'datetime',
-            'password' => 'hashed',
-        ];
-    }
-
-    public function sendPasswordResetNotification($token): void
-    {
-        $this->notify(new PasswordReset($token));
-    }
+    
 }
